refactor: rename Attachment to AttachmentChip and export it
(cherry picked from commit fd72d6dd1828a72b7b8177f70c8884bbc673ed0c)
